Fixes related to the change in txfee and minimum gas price, also some refactoring of duplicated code
