+++ qa/rpc-tests/p2p-compactblocks.py
@@ -123,8 +123,8 @@ class CompactBlocksTest(BitcoinTestFramework):
 
         # Start up node0 to be a version 1, pre-segwit node.
         self.nodes = start_nodes(self.num_nodes, self.options.tmpdir, 
-                [["-debug", "-logtimemicros=1", "-bip9params=segwit:0:0"], 
-                 ["-debug", "-logtimemicros", "-txindex"]])
+                [["-debug", "-logtimemicros=1", "-bip9params=segwit:0:0", "-minrelaytxfee=0.000001"], 
+                 ["-debug", "-logtimemicros", "-txindex", "-minrelaytxfee=0.000001"]])
         connect_nodes(self.nodes[0], 1)
 
     def build_block_on_tip(self, node, segwit=False):
