+++ qa/rpc-tests/qtum-dgp-gas-price-lingering-mempool-tx.py
@@ -8,21 +8,6 @@ from test_framework.qtum import *
 from test_framework.address import *
 from test_framework.blocktools import *
 
-class DGPState:
-    def __init__(self, node, contract_address):
-        self.node = node
-        self.contract_address = contract_address
-        self.abiAddAddressProposal = "bf5f1e83" #addAddressProposal(address,uint256)
-        self.abiSetInitialAdmin = "6fb81cbb" # setInitialAdmin()
-
-    def send_set_initial_admin(self, sender):
-        self.node.sendtoaddress(sender, 1)
-        self.node.sendtocontract(self.contract_address, self.abiSetInitialAdmin, 0, 4000000, 0.0001, sender)
-
-    def send_add_address_proposal(self, proposal_address, type1, sender):
-        self.node.sendtoaddress(sender, 1)
-        self.node.sendtocontract(self.contract_address, self.abiAddAddressProposal + proposal_address.zfill(64) + hex(type1)[2:].zfill(64), 0, 2000000, 0.001, sender)
-
 
 """
 Note, these tests do not test the functionality of the DGP template contract itself, for tests for the DGP template, see qtum-dgp.py
@@ -56,8 +41,8 @@ class QtumDGPGasPriceLingeringMempoolTxTest(BitcoinTestFramework):
         if min_gas_price == 1:
             contract_data = self.node.createcontract("6060604052602060405190810160405280600160ff168152506000906001610028929190610039565b50341561003457600080fd5b61010c565b8260016007016008900481019282156100c85791602002820160005b8382111561009657835183826101000a81548163ffffffff021916908360ff1602179055509260200192600401602081600301049283019260010302610055565b80156100c65782816101000a81549063ffffffff0219169055600401602081600301049283019260010302610096565b505b5090506100d591906100d9565b5090565b61010991905b8082111561010557600081816101000a81549063ffffffff0219169055506001016100df565b5090565b90565b6101628061011b6000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680633fb588191461003e575b600080fd5b341561004957600080fd5b610051610090565b6040518082600160200280838360005b8381101561007d5780820151818401525b602081019050610061565b5050505090500191505060405180910390f35b610098610108565b60006001806020026040519081016040528092919082600180156100fd576020028201916000905b82829054906101000a900463ffffffff1663ffffffff16815260200190600401906020826003010492830192600103820291508084116100c05790505b505050505090505b90565b6020604051908101604052806001905b600063ffffffff1681526020019060019003908161011857905050905600a165627a7a72305820485c90d0bd56d7e4a7c59cc8bfbe0a02b9c7106f74eaf66f03dc8b8b38b732030029", 1000000, 0.001)
 
-        if min_gas_price == 10:
-            contract_data = self.node.createcontract("6060604052602060405190810160405280600a60ff168152506000906001610028929190610039565b50341561003457600080fd5b61010c565b8260016007016008900481019282156100c85791602002820160005b8382111561009657835183826101000a81548163ffffffff021916908360ff1602179055509260200192600401602081600301049283019260010302610055565b80156100c65782816101000a81549063ffffffff0219169055600401602081600301049283019260010302610096565b505b5090506100d591906100d9565b5090565b61010991905b8082111561010557600081816101000a81549063ffffffff0219169055506001016100df565b5090565b90565b6101628061011b6000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680633fb588191461003e575b600080fd5b341561004957600080fd5b610051610090565b6040518082600160200280838360005b8381101561007d5780820151818401525b602081019050610061565b5050505090500191505060405180910390f35b610098610108565b60006001806020026040519081016040528092919082600180156100fd576020028201916000905b82829054906101000a900463ffffffff1663ffffffff16815260200190600401906020826003010492830192600103820291508084116100c05790505b505050505090505b90565b6020604051908101604052806001905b600063ffffffff1681526020019060019003908161011857905050905600a165627a7a7230582063e0364c58a4fdca79a0bf61b8b3dba16d95cdcba9bdab7997ef0a49879883b80029", 1000000, 0.001)
+        if min_gas_price == 100:
+            contract_data = self.node.createcontract("6060604052602060405190810160405280606460ff168152506000906001610028929190610039565b50341561003457600080fd5b61010c565b8260016007016008900481019282156100c85791602002820160005b8382111561009657835183826101000a81548163ffffffff021916908360ff1602179055509260200192600401602081600301049283019260010302610055565b80156100c65782816101000a81549063ffffffff0219169055600401602081600301049283019260010302610096565b505b5090506100d591906100d9565b5090565b61010991905b8082111561010557600081816101000a81549063ffffffff0219169055506001016100df565b5090565b90565b6101628061011b6000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680633fb588191461003e575b600080fd5b341561004957600080fd5b610051610090565b6040518082600160200280838360005b8381101561007d5780820151818401525b602081019050610061565b5050505090500191505060405180910390f35b610098610108565b60006001806020026040519081016040528092919082600180156100fd576020028201916000905b82829054906101000a900463ffffffff1663ffffffff16815260200190600401906020826003010492830192600103820291508084116100c05790505b505050505090505b90565b6020604051908101604052806001905b600063ffffffff1681526020019060019003908161011857905050905600a165627a7a72305820f1da304bd8da0e7c9401164a42f2b176afcfb4698bbf9aece7185d91fc5312720029", 1000000, 0.001)
 
         self.proposal_address = contract_data['address']
 
@@ -78,10 +63,10 @@ class QtumDGPGasPriceLingeringMempoolTxTest(BitcoinTestFramework):
         # until they connect later on
         stop_node(self.nodes[1], 1)
         self.nodes[1] = start_node(1, self.options.tmpdir)
-        self.nodes[1].createcontract("00", 1000000, 0.00000001)
+        self.nodes[1].createcontract("00", 1000000, QTUM_MIN_GAS_PRICE/COIN)
 
-        # Set the minimum gas price to 10
-        self.create_proposal_contract(10)
+        # Set the minimum gas price to 100
+        self.create_proposal_contract(100)
         self.GAS_PRICE_DGP.send_add_address_proposal(self.proposal_address, 2, admin_address)
         self.node.generate(2) # Activate the proposal
