+++ qa/rpc-tests/qtum-gas-limit.py
@@ -37,7 +37,7 @@ class QtumGasLimit(BitcoinTestFramework):
         # Create a tx with 2000 outputs each with a gas stipend of 5*10^8 calling the contract.
         tx = CTransaction()
         tx.vin = [make_vin(self.node, NUM_OUTPUTS*5*COIN)]
-        tx.vout = [CTxOut(0, CScript([1, 5*COIN, 1, b"\x00", bytes.fromhex(contract_address), OP_CALL])) for i in range(NUM_OUTPUTS)]
+        tx.vout = [CTxOut(0, CScript([b"\x04", int(5*COIN), QTUM_MIN_GAS_PRICE, b"\x00", bytes.fromhex(contract_address), OP_CALL])) for i in range(NUM_OUTPUTS)]
         tx.rehash()
         signed_tx_hex = self.node.signrawtransaction(bytes_to_hex_str(tx.serialize()))['hex']
