+++ qa/rpc-tests/qtum-identical-refunds.py
@@ -34,9 +34,9 @@ class QtumIdenticalRefunds(BitcoinTestFramework):
         # This will result in the same amount of gas being spent and thus the same amount being refunded.
         sender_address = self.node.getnewaddress()
         self.node.sendtoaddress(sender_address, 1)
-        self.node.sendtocontract(contract_address, "00", 0, 1000000, 0.00000001, sender_address)
+        self.node.sendtocontract(contract_address, "00", 0, 1000000, QTUM_MIN_GAS_PRICE/COIN, sender_address)
         self.node.sendtoaddress(sender_address, 1)
-        self.node.sendtocontract(contract_address, "00", 0, 1000000, 0.00000001, sender_address)
+        self.node.sendtocontract(contract_address, "00", 0, 1000000, QTUM_MIN_GAS_PRICE/COIN, sender_address)
 
         # Check that all txs were accepted into the mempool.
         assert_equal(len(self.node.getrawmempool()), 4)
