+++ qa/rpc-tests/qtum-many-value-refunds-from-same-tx.py
@@ -33,8 +33,8 @@ class QtumManyValueRefundsFromSameTxTest(BitcoinTestFramework):
         tx = CTransaction()
         tx.vin = [make_vin(self.node, int(20000*COIN))]
         tx.vout = []
-        tx.vout.append(CTxOut(int(COIN), CScript([b"\x04", CScriptNum(100000), CScriptNum(1), hex_str_to_bytes("00"), hex_str_to_bytes(contract_address), OP_CALL])))
-        tx.vout.append(CTxOut(int(COIN), CScript([b"\x04", CScriptNum(100000), CScriptNum(1), hex_str_to_bytes("00"), hex_str_to_bytes(contract_address), OP_CALL])))
+        tx.vout.append(CTxOut(int(COIN), CScript([b"\x04", CScriptNum(100000), CScriptNum(QTUM_MIN_GAS_PRICE), hex_str_to_bytes("00"), hex_str_to_bytes(contract_address), OP_CALL])))
+        tx.vout.append(CTxOut(int(COIN), CScript([b"\x04", CScriptNum(100000), CScriptNum(QTUM_MIN_GAS_PRICE), hex_str_to_bytes("00"), hex_str_to_bytes(contract_address), OP_CALL])))
 
         signed_tx_raw = self.node.signrawtransaction(bytes_to_hex_str(tx.serialize()))['hex']
         self.node.sendrawtransaction(signed_tx_raw)
