+++ qa/rpc-tests/qtum-opcreate.py
@@ -7,43 +7,8 @@ from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import *
 from test_framework.script import *
 from test_framework.mininode import *
+from test_framework.qtum import *
 import sys
-
-def make_vin(node, value):
-    addr = node.getrawchangeaddress()
-    txid_hex = node.sendtoaddress(addr, value/COIN)
-    txid = int(txid_hex, 16)
-    node.generate(1)
-    raw_tx = node.getrawtransaction(txid_hex, 1)
-
-    for vout_index, txout in enumerate(raw_tx['vout']):
-        #print i, txout['scriptPubKey']['addresses']
-        if txout['scriptPubKey']['addresses'] == [addr]:
-            break
-    else:
-        assert False
-
-    return CTxIn(COutPoint(txid, vout_index), nSequence=0)
-
-def make_op_create_output(node, value, version, gas_limit, gas_price, data):
-    scriptPubKey = CScript()
-    scriptPubKey += version
-    scriptPubKey += gas_limit
-    scriptPubKey += gas_price
-    scriptPubKey += data
-    scriptPubKey += OP_CREATE
-    return CTxOut(value, scriptPubKey)
-
-def make_op_create_transaction(node, vin, vout):
-    tx = CTransaction()
-    tx.vin = vin
-    tx.vout = vout
-    tx.rehash()
-    
-    unsigned_raw_tx = bytes_to_hex_str(tx.serialize_without_witness())
-    signed_raw_tx = node.signrawtransaction(unsigned_raw_tx)['hex']
-    return signed_raw_tx
-
 
 class OpCreateTest(BitcoinTestFramework):
     def __init__(self):
@@ -69,9 +34,9 @@ class OpCreateTest(BitcoinTestFramework):
             function () payable {}
         }
         """
-        tx = make_op_create_transaction(node,
+        tx = make_transaction(node,
             [self.vins.pop(-1)],
-            [make_op_create_output(node, 0, 4, 500000, b"\x0a", bytes.fromhex("60606040523415600b57fe5b5b60398060196000396000f30060606040525b600b5b5b565b0000a165627a7a72305820e3bed070fd3a81dd00e02efd22d18a3b47b70860155d6063e47e1e2674fc5acb0029"))]
+            [make_op_create_output(node, 0, 4, CScriptNum(500000), CScriptNum(QTUM_MIN_GAS_PRICE), bytes.fromhex("60606040523415600b57fe5b5b60398060196000396000f30060606040525b600b5b5b565b0000a165627a7a72305820e3bed070fd3a81dd00e02efd22d18a3b47b70860155d6063e47e1e2674fc5acb0029"))]
         )
         #node.createcontract("60606040523415600b57fe5b5b60398060196000396000f30060606040525b600b5b5b565b0000a165627a7a72305820e3bed070fd3a81dd00e02efd22d18a3b47b70860155d6063e47e1e2674fc5acb0029");
         node.sendrawtransaction(tx)
@@ -104,7 +69,7 @@ class OpCreateTest(BitcoinTestFramework):
             function () payable {}
         }
         """
-        node.createcontract("606060405234610000575b61034a806100196000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680633f811b80146100495780636b8ff5741461006a575b610000565b3461000057610068600480803560001916906020019091905050610087565b005b3461000057610085600480803590602001909190505061015b565b005b60008160405160e18061023e833901808260001916600019168152602001915050604051809103906000f08015610000579050600180548060010182818154818355818115116101035781836000526020600020918201910161010291905b808211156100fe5760008160009055506001016100e6565b5090565b5b505050916000526020600020900160005b83909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550505b5050565b6000600182815481101561000057906000526020600020900160005b9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690508073ffffffffffffffffffffffffffffffffffffffff16638052474d6000604051602001526040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401809050602060405180830381600087803b156100005760325a03f1156100005750505060405180519050600083815481101561000057906000526020600020900160005b5081600019169055505b50505600606060405234610000576040516020806100e1833981016040528080519060200190919050505b80600081600019169055505b505b609f806100426000396000f30060606040523615603d576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680638052474d146045575b60435b5b565b005b34600057604f606d565b60405180826000191660001916815260200191505060405180910390f35b600054815600a165627a7a723058209bd85f2ac8e941766991a59f8a0183902c8168f671bfdb430ad9eab85fd697b70029a165627a7a72305820ed128e4929006fce038bc859dd8837890e7ee8d296cd3ed30b66603a8423397e0029", 1000000, 0.00000001)
+        node.createcontract("606060405234610000575b61034a806100196000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680633f811b80146100495780636b8ff5741461006a575b610000565b3461000057610068600480803560001916906020019091905050610087565b005b3461000057610085600480803590602001909190505061015b565b005b60008160405160e18061023e833901808260001916600019168152602001915050604051809103906000f08015610000579050600180548060010182818154818355818115116101035781836000526020600020918201910161010291905b808211156100fe5760008160009055506001016100e6565b5090565b5b505050916000526020600020900160005b83909190916101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550505b5050565b6000600182815481101561000057906000526020600020900160005b9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690508073ffffffffffffffffffffffffffffffffffffffff16638052474d6000604051602001526040518163ffffffff167c0100000000000000000000000000000000000000000000000000000000028152600401809050602060405180830381600087803b156100005760325a03f1156100005750505060405180519050600083815481101561000057906000526020600020900160005b5081600019169055505b50505600606060405234610000576040516020806100e1833981016040528080519060200190919050505b80600081600019169055505b505b609f806100426000396000f30060606040523615603d576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680638052474d146045575b60435b5b565b005b34600057604f606d565b60405180826000191660001916815260200191505060405180910390f35b600054815600a165627a7a723058209bd85f2ac8e941766991a59f8a0183902c8168f671bfdb430ad9eab85fd697b70029a165627a7a72305820ed128e4929006fce038bc859dd8837890e7ee8d296cd3ed30b66603a8423397e0029", 1000000, 0.000001)
         block_height = node.getblockcount()
         node.generate(1)
         sync_blocks(self.nodes)
@@ -162,7 +127,7 @@ class OpCreateTest(BitcoinTestFramework):
             function () payable {}
         }
         """
-        tx = make_op_create_transaction(node,
+        tx = make_transaction(node,
             [self.vins.pop(-1)],
             # changing the gas limit \xff\xff -> \xff\xff\x00 results in success.
             [make_op_create_output(node, 0, b"\x04", b"\xff\xff", 1000, bytes.fromhex("60606040523415600b57fe5b5b60398060196000396000f30060606040525b600b5b5b565b0000a165627a7a7230582092926a9814888ff08700cbd86cf4ff8c50052f5fd894e794570d9551733591d60029"))]
@@ -184,7 +149,7 @@ class OpCreateTest(BitcoinTestFramework):
             function () payable {}
         }
         """
-        tx = make_op_create_transaction(node,
+        tx = make_transaction(node,
             [self.vins.pop(-1)],
             # changing the gas limit \xff\xff -> \xff\xff\x00 results in success.
             [make_op_create_output(node, 0, b"\x04", b"\xff\x4f", 1000, bytes.fromhex("60606040523415600b57fe5b5b60398060196000396000f30060606040525b600b5b5b565b0000a165627a7a7230582092926a9814888ff08700cbd86cf4ff8c50052f5fd894e794570d9551733591d60029")),
