+++ qa/rpc-tests/qtum-sendtocontract.py
@@ -7,6 +7,7 @@ from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import *
 from test_framework.script import *
 from test_framework.mininode import *
+from test_framework.qtum import *
 import sys
 import time
 
@@ -44,7 +45,7 @@ class SendToContractTest(BitcoinTestFramework):
             }
         }
         """
-        contract_data = self.node.createcontract("6060604052341561000c57fe5b5b33600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b5b6102218061005f6000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806313af40351461005c5780632bcf51b41461008a5780635e01eb5a14610094578063893d20e8146100e6575bfe5b610088600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610138565b005b61009261017d565b005b341561009c57fe5b6100a46101c1565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34156100ee57fe5b6100f66101ca565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b80600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b50565b33600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b565b60003390505b90565b6000600060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690505b905600a165627a7a72305820791895230daae0ed58cc374ad5b639044408f1942d7b85689a616caee50dc42e0029", 1000000, 0.00000001)
+        contract_data = self.node.createcontract("6060604052341561000c57fe5b5b33600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b5b6102218061005f6000396000f30060606040526000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806313af40351461005c5780632bcf51b41461008a5780635e01eb5a14610094578063893d20e8146100e6575bfe5b610088600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610138565b005b61009261017d565b005b341561009c57fe5b6100a46101c1565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34156100ee57fe5b6100f66101ca565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b80600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b50565b33600060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505b565b60003390505b90565b6000600060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690505b905600a165627a7a72305820791895230daae0ed58cc374ad5b639044408f1942d7b85689a616caee50dc42e0029", 1000000, QTUM_MIN_GAS_PRICE/COIN)
         self.contract_address = contract_data['address']
         self.node.generate(1)
 
@@ -87,7 +88,7 @@ class SendToContractTest(BitcoinTestFramework):
         self.node.sendtoaddress("qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT", 0.1)
         self.node.generate(1)
         # call setSenderAsOwner with 100 qtum
-        ret = self.node.sendtocontract(self.contract_address, "2bcf51b4", 100, 1000000, 0.00000001, "qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT")
+        ret = self.node.sendtocontract(self.contract_address, "2bcf51b4", 100, 1000000, QTUM_MIN_GAS_PRICE/COIN, "qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT")
         assert('txid' in ret)
         assert('sender' in ret)
         assert('hash160' in ret)
@@ -105,7 +106,7 @@ class SendToContractTest(BitcoinTestFramework):
         self.node.sendtoaddress("qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT", 0.1)
         self.node.generate(1)
         # call setSenderAsOwner with 100 qtum
-        ret = self.node.sendtocontract(self.contract_address, "2bcf51b4", 100, 1000000, 0.00000001, "qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT", False)
+        ret = self.node.sendtocontract(self.contract_address, "2bcf51b4", 100, 1000000, QTUM_MIN_GAS_PRICE/COIN, "qabmqZk3re5b9UpUcznxDkCnCsnKdmPktT", False)
         assert('raw transaction' in ret)
         assert(len(ret.keys()) == 1)
         decoded_tx = self.node.decoderawtransaction(ret['raw transaction'])
@@ -116,7 +117,7 @@ class SendToContractTest(BitcoinTestFramework):
         assert(False)
 
     def run_test(self):
-        self.nodes[0].generate(COINBASE_MATURITY+100)
+        self.nodes[0].generate(100+COINBASE_MATURITY)
         self.setup_contract()
         self.sendtocontract_verify_storage_test()
         self.sendtocontract_verify_storage_and_balance_test()
